Lossy conversion when casting double to int.

+++ src/main/java/Main.java
@@ -3,7 +3,7 @@ public class Main {
     // Type conversion: double to integer
     
     double number1 = 5.8;
-    int number2 = number1;
+    int number2 = (int) number1;
 
     // Incompatible types: possible lossy conversion from double to int
     System.out.println("Number 1 is " + number1);
